+++ packages/qr_code_dart_decoder/lib/qr_code_dart_decoder.dart
@@ -9,6 +9,7 @@ import 'package:qr_code_dart_decoder/src/camera/yuv420_planes.dart';
 import 'package:qr_code_dart_decoder/src/file/file_decode.dart';
 import 'package:qr_code_dart_decoder/src/file/file_decode_event.dart';
 import 'package:qr_code_dart_decoder/src/util/crop_rect.dart';
+import 'package:qr_code_dart_decoder/src/util/rotation_type.dart';
 import 'package:zxing_lib/zxing.dart';
 
 export 'package:qr_code_dart_decoder/src/camera/camera_decode.dart';
@@ -17,6 +18,7 @@ export 'package:qr_code_dart_decoder/src/camera/yuv420_planes.dart';
 export 'package:qr_code_dart_decoder/src/file/file_decode.dart';
 export 'package:qr_code_dart_decoder/src/file/file_decode_event.dart';
 export 'package:qr_code_dart_decoder/src/util/crop_rect.dart';
+export 'package:qr_code_dart_decoder/src/util/rotation_type.dart';
 export 'package:zxing_lib/zxing.dart' show BarcodeFormat, Result;
 
 /// A Calculator.
@@ -42,7 +44,7 @@ class QrCodeDartDecoder {
   Future<Result?> decodeFile(
     Uint8List bytes, {
     bool isInverted = false,
-    bool rotate = false,
+    RotationType? rotate,
     CropRect? cropRect,
   }) async {
     final image = decodeImage(bytes);
@@ -52,7 +54,7 @@ class QrCodeDartDecoder {
       formats: formats,
       width: image.width,
       height: image.height,
-      rotate: rotate,
+      rotation: rotate,
       cropRect: cropRect,
     );
     return FileDecode.decode(event.toMap());
@@ -61,14 +63,14 @@ class QrCodeDartDecoder {
   Future<Result?> decodeCameraImage(
     List<Yuv420Planes> yuv420Planes, {
     bool isInverted = false,
-    bool rotate = false,
+    RotationType? rotate,
     CropRect? cropRect,
   }) async {
     final event = CameraDecodeEvent(
       yuv420Planes: yuv420Planes,
       invert: isInverted,
       formats: formats,
-      rotate: rotate,
+      rotation: rotate,
       cropRect: cropRect,
     );
     return CameraDecode.decode(event.toMap());