+++ packages/qr_code_dart_decoder/lib/src/camera/camera_decode_event.dart
@@ -1,11 +1,12 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:qr_code_dart_decoder/src/camera/yuv420_planes.dart';
 import 'package:qr_code_dart_decoder/src/util/crop_rect.dart';
+import 'package:qr_code_dart_decoder/src/util/rotation_type.dart';
 import 'package:zxing_lib/zxing.dart';
 
 class CameraDecodeEvent {
   final bool invert;
-  final bool rotate;
+  final RotationType? rotation;
   final List<Yuv420Planes> yuv420Planes;
   final List<BarcodeFormat> formats;
   final CropRect? cropRect;
@@ -13,21 +14,21 @@ class CameraDecodeEvent {
   CameraDecodeEvent({
     required this.yuv420Planes,
     this.invert = false,
-    this.rotate = false,
+    this.rotation,
     this.formats = const [],
     this.cropRect,
   });
 
   CameraDecodeEvent copyWith({
     bool? invert,
-    bool? rotate,
+    RotationType? rotationType,
     List<Yuv420Planes>? yuv420Planes,
     List<BarcodeFormat>? formats,
     CropRect? cropRect,
   }) {
     return CameraDecodeEvent(
       invert: invert ?? this.invert,
-      rotate: rotate ?? this.rotate,
+      rotation: rotationType ?? rotation,
       yuv420Planes: yuv420Planes ?? this.yuv420Planes,
       formats: formats ?? this.formats,
       cropRect: cropRect ?? this.cropRect,
@@ -37,7 +38,7 @@ class CameraDecodeEvent {
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'invert': invert,
-      'rotate': rotate,
+      'rotation': rotation?.name,
       'yuv420Planes': yuv420Planes.map((e) => e.toMap()).toList(),
       'formats': formats.map((e) => e.name).toList(),
       'cropRect': cropRect?.toMap(),
@@ -47,7 +48,7 @@ class CameraDecodeEvent {
   factory CameraDecodeEvent.fromMap(Map<String, dynamic> map) {
     return CameraDecodeEvent(
       invert: map['invert'] as bool,
-      rotate: map['rotate'] as bool,
+      rotation: map['rotation'] != null ? RotationType.values.byName(map['rotation']) : null,
       yuv420Planes: (map['yuv420Planes'] as List)
           .map((e) => Yuv420Planes.fromMap((e as Map).cast()))
           .toList(),