+++ packages/qr_code_dart_decoder/lib/src/file/file_decode.dart
@@ -16,7 +16,7 @@ abstract class FileDecode {
         imageBytes,
         event.width,
         event.height,
-        rotateCounterClockwise: event.rotate,
+        rotationType: event.rotation,
         cropRect: event.cropRect,
       );